BankingNLP_v4/requirements.txt
########################################
# Модули для анализа данных и ML
########################################
pandas>=2.0.0              # Работа с таблицами и данными
numpy>=1.25.0               # Математика и массивы
scikit-learn>=1.3.0        # Классические ML-модели
matplotlib>=3.7.0          # Визуализация графиков
seaborn>=0.13.0             # Красивые графики
joblib>=1.3.0              # Сериализация моделей
psutil>=5.9.0              # Мониторинг ресурсов

########################################
# Веб-фреймворк и сервер
########################################
fastapi>=0.110.0             # Основной backend-фреймворк
uvicorn>=0.25.0             # ASGI-сервер для FastAPI
jinja2>=3.1.0              # Шаблонизатор для HTML
python-multipart>=0.0.7    # Загрузка файлов через формы

########################################
# Pydantic и валидация
########################################
pydantic>=2.5.0            # Валидация данных
pydantic-settings>=2.1.0   # Настройки через Pydantic

########################################
# Безопасность и авторизация
########################################
python-jose>=3.3.0         # JWT-токены
passlib[bcrypt]>=1.7.4     # Хеширование паролей

########################################
# Работа с HTTP и API
########################################
requests>=2.31.0            # Синхронные HTTP-запросы
aiohttp>=3.9.0             # Асинхронные HTTP-запросы

########################################
# Работа с YAML и конфигами
########################################
pyyaml>=6.0              # Загрузка и сохранение YAML

########################################
# Логирование
########################################
loguru>=0.7.2              # Современное логирование

########################################
# LLM/Inference (малые языковые модели)
########################################
llama-cpp-python>=0.2.70    # Инференс локальных LLM (Phi-2, Llama и др.)

########################################
# Работа с базой данных
########################################
sqlalchemy>=2.0.0          # ORM для работы с БД

########################################
# Тестирование
########################################
pytest>=7.4.0              # Фреймворк для тестов
httpx>=0.27.0               # HTTP-клиент для тестирования API

########################################
# Explainability (опционально)
########################################
shap>=0.44.0                # Интерпретируемость моделей

########################################
# Инструменты для разработки (dev)
########################################
black>=23.0.0               # Форматирование кода
isort>=5.12.0               # Сортировка импортов
flake8>=6.0.0              # Линтер

########################################
# Русский язык и NLP
########################################
spacy>=3.7.0               # NLP-библиотека

########################################
# Сервер для продакшена
########################################
gunicorn>=21.2.0            # WSGI/ASGI сервер (используется в Docker) 